Application/requirements.txt
#-f https://download.pytorch.org/whl/torch_stable.html
torch==1.2.0 
torchsummary==1.5.1
pytorch-lightning==0.7.3 
torchvision==0.4.0  
pytorch-toolbelt==0.2.1 
flask==1.1.2
flask-basicauth==0.2.0
mgrs==1.3.6
opencv-python==4.2.0.32 
pillow==5.3.0
utm==0.5.0
bokeh==2.0.2 
matplotlib==3.2.1
neptune-cli
neptune-client
timm==0.1.14
albumentations==0.4.5
glob2==0.6
numpy==1.16.4

#otherstuff
itsdangerous
MarkupSafe==1.0